Update time granularity solver to use PTD.
The time granularity solver previously relied on the model to determine the
valid time granularities for time dimensions. With the introduction of the
plot time dimension, the time granularity solver needs to apply the same
transformation to that the dataflow plan builder runs to ensure that it can
resolve queries for "valid granularities for this metric and the plot time
dimension".